+++ src/components/TableHeadCell.js
@@ -8,6 +8,7 @@ import PropTypes from 'prop-types';
 import React, {useState,useEffect} from 'react';
 import MuiTooltip from '@material-ui/core/Tooltip';
 import {useDrop, useDrag} from 'react-dnd';
+import useColumnDrop from '../hooks/useColumnDrop.js';
 
 const useStyles = makeStyles(theme => ({
   root: {},
@@ -81,7 +82,6 @@ function TableHeadCell(props) {
     index,
     print,
     column,
-    reorderColumns,
     updateColumnOrder,
     columnOrder,
     setCellRef,
@@ -91,7 +91,7 @@ function TableHeadCell(props) {
     timers,
     components = {},
   } = props;
-  const [dragging, setDragging] = props.draggingHook;
+  const [dragging, setDragging] = props.draggingHook ? props.draggingHook : [];
 
   const { className, ...otherProps } = cellHeaderProps;
   const Tooltip = components.Tooltip || MuiTooltip;
@@ -126,113 +126,25 @@ function TableHeadCell(props) {
     }
   });
 
-  function getColModel(headCellRefs, columnOrder) {
-    var colModel = [];
-
-    var ii = 0,
-      parentOffsetLeft = 0,
-      offsetParent = headCellRefs[0].offsetParent;
-    while (offsetParent){
-      parentOffsetLeft = parentOffsetLeft + (offsetParent.offsetLeft || 0);
-      offsetParent = offsetParent.offsetParent;
-      ii++;
-      if (ii > 1000) {
-        console.warn('Table nested within 1000 divs. Maybe an error.');
-        break;
-      }
-    }
-
-    colModel[0] = {
-      left: parentOffsetLeft + headCellRefs[0].offsetLeft,
-      width: headCellRefs[0].offsetWidth,
-      columnIndex: null,
-      ref: headCellRefs[0]
-    };
-
-    columnOrder.forEach( (colIdx, idx) => {
-      var col = headCellRefs[colIdx + 1];
-      var cmIndx = colModel.length - 1;
-      colModel.push({
-        left: colModel[cmIndx].left + colModel[cmIndx].width,
-        width: col.offsetWidth,
-        columnIndex: colIdx,
-        ref: col
-      });
-    });
-
-    return colModel;
-  }
-
-  const [{ isOver, canDrop }, drop] = useDrop({
-    accept: 'HEADER',
+  const [drop] = useColumnDrop({
     drop: (item, mon) => {
       setSortTooltipOpen(false);
       setHintTooltipOpen(false);
       setDragging(false);
     },
-    hover: (item, mon) => {
-
-      var hoverIdx = mon.getItem().colIndex;
-
-      if ( hoverIdx !== index ) {
-
-        var newColModel = getColModel( headCellRefs, reorderColumns( columnOrder, mon.getItem().colIndex, index ));
-
-        var newX = mon.getClientOffset().x;
-        var modelIdx = -1;
-        for (var ii = 0; ii < newColModel.length; ii++) {
-          if (newX > newColModel[ii].left && newX < (newColModel[ii].left + newColModel[ii].width) ) {
-            modelIdx = newColModel[ii].columnIndex;
-            break;
-          }
-        }
-
-        if (modelIdx === mon.getItem().colIndex ) {
-          clearTimeout( timers.columnShift );
-
-          var ttime = options.draggableColumns.transitionTime || 300;
-          var curColModel = getColModel( headCellRefs,  columnOrder );
-          
-          var transitions = [];
-          newColModel.forEach(item => {
-            transitions[item.columnIndex] = item.left;
-          });
-          curColModel.forEach(item => {
-            transitions[item.columnIndex] = transitions[item.columnIndex] - item.left;
-          });
-
-          for (var idx = 1; idx < columnOrder.length; idx++) {
-            headCellRefs[idx].style.transition = '280ms';
-            headCellRefs[idx].style.transform = 'translateX(' + transitions[idx-1] + 'px)';
-          };
-
-          var allElms = [];
-          for (var ii = 0; ii < columnOrder.length; ii++) { 
-            var table = tableRef();
-            var elms = table ? table.querySelectorAll('[data-colindex="' + ii + '"]') : [];
-            for (var jj = 0; jj < elms.length; jj++) {
-              elms[jj].style.transition = ttime + 'ms';
-              elms[jj].style.transform = 'translateX(' + transitions[ii] + 'px)';
-              allElms.push( elms[jj] );
-            }
-          }
-
-          var newColIndex = mon.getItem().colIndex;
-          timers.columnShift = setTimeout(() => {
-            allElms.forEach( item => {
-              item.style.transition = '0s';
-              item.style.transform = 'translateX(0)';
-            });
-            updateColumnOrder( columnOrder, newColIndex, index );
-          }, ttime);
-        }
-      }
-    },
-    collect: mon => ({
-      isOver: !!mon.isOver(),
-      canDrop: !!mon.canDrop(),
-    }),
+    index,
+    headCellRefs,
+    updateColumnOrder,
+    columnOrder,
+    transitionTime: options.draggableColumns.transitionTime,
+    tableRef: tableRef(),
+    timers
   });
+
+  const isDraggingEnabled = () => {
+    if (!props.draggingHook) return false;
+    return options.draggableColumns.enabled && column.draggable !== false;
+  };
 
   const cellClass = classNames({
     [classes.root]: true,
@@ -249,7 +161,7 @@ function TableHeadCell(props) {
   let refProp = {};
   refProp.ref = el => {
     drop(el);
-    setCellRef && setCellRef(index + 1, findDOMNode(el));
+    setCellRef && setCellRef(index + 1, el);
   };
   
   const getTooltipTitle = () => {
@@ -273,7 +185,7 @@ function TableHeadCell(props) {
           onClick={handleSortClick}
           className={classes.toolButton}
           data-testid={'headcol-' + props.index}
-          ref={(options.draggableColumns.enabled && column.draggable !== false) ? dragRef : null}
+          ref={isDraggingEnabled() ? dragRef : null}
           tabIndex={0}>
           <Tooltip
             title={getTooltipTitle()}
@@ -308,7 +220,7 @@ function TableHeadCell(props) {
           )}
         </span>
       ) : (
-        <div className={hint ? classes.sortAction : null} ref={(options.draggableColumns.enabled && column.draggable !== false) ? dragRef : null}>
+        <div className={hint ? classes.sortAction : null} ref={isDraggingEnabled() ? dragRef : null}>
           {children}
           {hint && (
             <Tooltip
